DEV: `@oldValue` in modal changes due to `.focus()` before exec
Previously we were passing `this.selectedText` to `@oldValue` of the `DiffModal` however, `this.selectedText` was being set to `""` because of the `.focus()` happening before `execCommand()`. So instead, we store the initial value in its own tracked property only when the AI request is made so we can pass it to the modal without it being altered.