FEATURE: Share conversations with AI via a URL
This allows users to share a static page of an AI conversation with
the rest of the world.

A follow on will be a custom onebox in Discourse to display this
nicely.

By default this feature is disabled, it is enabled by turning on
ai_bot_allow_public_sharing via site settings

Precautions are taken when sharing

1. We make a carbonite copy
2. We minimize work generating page
3. We limit to 100 interactions
4. Many security checks - including disallowing if there is a mix
of users in the PM.