UX: Add title suffix to shared AI pages
Previously, the title of the shared AI page would be the title of the PM. When shared somewhere else (e.g. another Discourse site, or social media platform), this can make it look like an authoritative article published by the forum owner.

This commit adds an `- AI Conversation` suffix to the page title.

This serves a few purposes:

1. Increases awareness of discourse-ai's bot feature

2. Bring external sharing in line with the internal-onebox for this feature, which already has a clear 'AI conversation' label

3. Makes it clearer that these pages are not an authoritative source, and are not necessarily endorsed by the forum owner

This is especially important in situations where the og:description is not displayed (e.g. in Discourse inline oneboxes)