FEATURE: allow llm triage to automatically hide posts
Previous to this change we could flag, but there was no way
to hide content.

We had the option to hide topics, but this is not desirable for
a spam reply.

New option allows triage to hide a post if it is a reply, if the
post happens to be the first post on the topic, the topic will
be hidden.