FEATURE: AI artifacts
Initial implementation of an artifact system which allows users
to generate HTML pages directly from the AI persona.

FEATURE: support tool progress callbacks

This is anthropic only for now, but we can get a callback as tool
is completing, this gives us the ability to show progress to user
as the function is populating.

work in progress

Revert "work in progress"

This reverts commit 30ebe562ea4e5601701ff89af80b964e1cc6af5e.

Revert "FEATURE: support tool progress callbacks"

This reverts commit fd7ccfd0ab538d2845b096d3902dc06bdbf5cf1f.