FIX: results not being reset when appending to query param
This PR fixes an issue where the AI search results were not being reset when you append your search to an existing query param (typically when you've come from quick search). This is because `handleSearch()` doesn't get called in this situation. So here we explicitly check for query params, trigger a reset and search for those occasions. Testing needs to be improved quite a bit here so I will address it in a follow-up PR.